backend/app/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
bidict==0.23.1
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
flasgger==0.9.7.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-OAuthlib==0.9.6
frozenlist==1.4.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwt==1.3.1
MarkupSafe==2.1.5
mistune==3.0.2
multidict==6.0.5
oauthlib==2.1.0
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
python-engineio==4.9.0
python-socketio==5.11.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.1.0
rpds-py==0.18.0
simple-websocket==1.0.0
six==1.16.0
urllib3==2.2.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
